[WEB-3706] - Made work for content update
